+++ CHANGELOG.md
@@ -13,8 +13,8 @@ CHANGELOG
   # Or you can just hide it like so
   fzf --preview 'cat {}' --preview-window '<50(hidden)'
   ```
-- Use SGR mouse mode to support larger terminals
-- You can now use characters that does not satisfy `unicode.IsGraphic` constraint
+- fzf now uses SGR mouse mode to properly support mouse on larger terminals
+- You can now use characters that do not satisfy `unicode.IsGraphic` constraint
   for `--marker`, `--pointer`, and `--ellipsis`. Allows Nerd Fonts and stuff.
   Use at your own risk.
 - Bug fixes and improvements